refactor(core): extract cadence streaming pipeline from http.rs
Move fixed-cadence audio stream, delivery tracking guard, and related
helpers into stream/cadence.rs. Consolidate duplicated silence tracking
into the cadence stream itself, replacing the 20-line external wrapper
with a single TaggedFrame::into_bytes unwrap.
