+++ packages/thaumic-core/src/api/http.rs
@@ -2,15 +2,12 @@
 //!
 //! All handlers are thin - they delegate to services for business logic.
 
-use std::collections::VecDeque;
 use std::net::{IpAddr, SocketAddr};
 use std::path::PathBuf;
 use std::pin::Pin;
-use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 
-use async_stream::stream;
 use axum::{
     body::Body,
     extract::{connect_info::ConnectInfo, Path, State},
@@ -23,8 +20,6 @@ use bytes::Bytes;
 use futures::stream::{Stream, StreamExt};
 use serde::Deserialize;
 use serde_json::json;
-use tokio::sync::broadcast;
-use tokio::time::{interval, Instant as TokioInstant, MissedTickBehavior};
 use tokio_stream::wrappers::errors::BroadcastStreamRecvError;
 use tokio_stream::wrappers::BroadcastStream;
 
@@ -39,395 +34,13 @@ use crate::protocol_constants::{
 use crate::sonos::discovery::probe_speaker_by_ip;
 use crate::state::ManualSpeakerConfig;
 use crate::stream::{
-    apply_fade_in, create_fade_out_frame, create_wav_header, crossfade_samples,
-    extract_last_sample_pair, is_crossfade_compatible, AudioCodec, AudioFormat,
-    IcyMetadataInjector, TaggedFrame,
+    create_wav_header, create_wav_stream_with_cadence, lagged_error, AudioCodec,
+    IcyMetadataInjector, LoggingStreamGuard, TaggedFrame,
 };
 use crate::utils::validate_speaker_ip;
 
 /// Boxed stream type for audio data with ICY metadata support.
 type AudioStream = Pin<Box<dyn futures::Stream<Item = Result<Bytes, std::io::Error>> + Send>>;
-
-/// Threshold for counting delivery gaps (100ms).
-/// PCM at 48kHz stereo 16-bit = 192KB/s, so 100ms = ~19KB of audio.
-const DELIVERY_GAP_THRESHOLD_MS: u64 = 100;
-
-/// Only log gaps exceeding this threshold to avoid log spam (500ms).
-const DELIVERY_GAP_LOG_THRESHOLD_MS: u64 = 500;
-
-/// Creates an IO error for broadcast channel lag.
-///
-/// Logs a warning and returns a formatted error. Centralizes the handling
-/// of `BroadcastStreamRecvError::Lagged` to avoid duplication.
-fn lagged_error(frames: u64) -> std::io::Error {
-    log::warn!(
-        "[Stream] Broadcast receiver lagged by {} frames - possible CPU contention",
-        frames
-    );
-    std::io::Error::other(format!("lagged by {} frames", frames))
-}
-
-/// Creates a WAV audio stream with fixed-cadence output and crossfade on silence transitions.
-///
-/// Maintains real-time cadence regardless of input timing:
-/// - Incoming frames are queued (bounded to `queue_size`)
-/// - Metronome ticks every `frame_duration_ms`
-/// - On each tick: send queued frame if available, else send silence
-///
-/// Crossfade on silence transitions:
-/// - When entering silence: emits a fade-out frame from the last audio sample to zero
-/// - When exiting silence: applies fade-in to the first audio frame
-///
-/// This ensures Sonos always receives continuous data with smooth transitions,
-/// eliminating pops from abrupt audio/silence boundaries.
-fn create_wav_stream_with_cadence(
-    mut rx: broadcast::Receiver<Bytes>,
-    silence_frame: Bytes,
-    guard: Arc<LoggingStreamGuard>,
-    queue_size: usize,
-    frame_duration_ms: u32,
-    audio_format: AudioFormat,
-    prefill_frames: Vec<Bytes>,
-) -> impl Stream<Item = Result<TaggedFrame, std::io::Error>> {
-    stream! {
-        let cadence_duration = Duration::from_millis(frame_duration_ms as u64);
-
-        // Pre-populate queue with prefill frames to eliminate handoff gap.
-        // This ensures the first tick immediately yields audio.
-        let mut queue: VecDeque<Bytes> = VecDeque::with_capacity(queue_size.max(prefill_frames.len()));
-        for frame in prefill_frames {
-            queue.push_back(frame);
-        }
-
-        // Fire first tick immediately to get audio flowing before Sonos times out.
-        // On resume, Sonos closes the connection within milliseconds if no audio arrives.
-        let mut metronome = interval(cadence_duration);
-        metronome.set_missed_tick_behavior(MissedTickBehavior::Burst);
-
-        let mut rx_closed = false;
-        let mut in_silence = false;
-        let mut silence_start: Option<TokioInstant> = None;
-
-        // Crossfade is only supported for 16-bit PCM. The handshake enforces this
-        // for PCM streams, but we guard here defensively in case the invariant breaks.
-        let crossfade_enabled = is_crossfade_compatible(&audio_format);
-        if !crossfade_enabled {
-            log::warn!(
-                "[Stream] Crossfade disabled: requires 16-bit PCM, got {}-bit",
-                audio_format.bits_per_sample
-            );
-        }
-
-        // Crossfade state: track last sample pair for fade-out generation
-        let fade_samples = crossfade_samples(audio_format.sample_rate);
-        let samples_per_frame = audio_format.frame_samples(frame_duration_ms);
-        let mut last_sample_pair: Option<(i16, i16)> = None;
-
-        // Rate-limit lagged warnings (max once per second)
-        let mut last_lagged_log: Option<TokioInstant> = None;
-
-        loop {
-            // Exit when channel closed AND queue drained
-            if rx_closed && queue.is_empty() {
-                break;
-            }
-
-            tokio::select! {
-                biased;
-
-                // PRIORITY 1: Metronome tick - MUST emit something every frame_duration_ms
-                _ = metronome.tick() => {
-                    if let Some(frame) = queue.pop_front() {
-                        // Real audio available
-                        let was_in_silence = in_silence;
-                        if in_silence {
-                            if let Some(start) = silence_start.take() {
-                                log::info!(
-                                    "[Stream] Exiting silence (cadence) after {:.1}s",
-                                    start.elapsed().as_secs_f32()
-                                );
-                            }
-                            in_silence = false;
-                        }
-
-                        // Track last sample pair for potential future fade-out
-                        if crossfade_enabled {
-                            last_sample_pair = extract_last_sample_pair(&frame, audio_format.channels);
-                        }
-
-                        // Apply fade-in if we just exited silence (and crossfade is enabled)
-                        if was_in_silence && crossfade_enabled {
-                            let mut faded = frame.to_vec();
-                            apply_fade_in(&mut faded, audio_format.channels, fade_samples);
-                            yield Ok(TaggedFrame::Audio(Bytes::from(faded)));
-                        } else {
-                            yield Ok(TaggedFrame::Audio(frame));
-                        }
-                    } else if !rx_closed {
-                        // No frame available, emit silence
-                        let entering_silence = !in_silence;
-                        if entering_silence {
-                            log::info!("[Stream] Entering silence (cadence) - queue empty");
-                            in_silence = true;
-                            silence_start = Some(TokioInstant::now());
-
-                            // Generate fade-out frame if crossfade is enabled and we have last sample values
-                            if crossfade_enabled {
-                                if let Some((left, right)) = last_sample_pair.take() {
-                                    let fade_out = create_fade_out_frame(
-                                        left,
-                                        right,
-                                        audio_format.channels,
-                                        fade_samples,
-                                        samples_per_frame,
-                                    );
-                                    yield Ok(TaggedFrame::Silence(fade_out));
-                                } else {
-                                    yield Ok(TaggedFrame::Silence(silence_frame.clone()));
-                                }
-                            } else {
-                                yield Ok(TaggedFrame::Silence(silence_frame.clone()));
-                            }
-                        } else {
-                            yield Ok(TaggedFrame::Silence(silence_frame.clone()));
-                        }
-                    }
-                    // If rx_closed and queue empty, don't yield - loop will break
-
-                    // Drain any pending frames from rx into queue after emitting.
-                    // This prevents starvation: with biased select, ticks always win,
-                    // so without this drain, frames could pile up in rx while we
-                    // emit silence (especially during recovery from underflow).
-                    if !rx_closed {
-                        loop {
-                            match rx.try_recv() {
-                                Ok(frame) => {
-                                    if queue.len() >= queue_size {
-                                        queue.pop_front();
-                                        guard.record_frame_dropped();
-                                    }
-                                    queue.push_back(frame);
-                                }
-                                Err(broadcast::error::TryRecvError::Empty) => break,
-                                Err(broadcast::error::TryRecvError::Lagged(n)) => {
-                                    let now = TokioInstant::now();
-                                    if last_lagged_log.map_or(true, |t| now.duration_since(t).as_secs() >= 1) {
-                                        log::warn!("[Stream] Lagged by {n} frames (during drain)");
-                                        last_lagged_log = Some(now);
-                                    }
-                                    // Continue draining - next try_recv will get latest
-                                }
-                                Err(broadcast::error::TryRecvError::Closed) => {
-                                    rx_closed = true;
-                                    log::debug!("[Stream] Channel closed, draining {} queued frames", queue.len());
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-
-                // PRIORITY 2: Receive frames into queue (when channel open and tick not ready)
-                result = rx.recv(), if !rx_closed => {
-                    match result {
-                        Ok(frame) => {
-                            if queue.len() >= queue_size {
-                                // Queue full - drop oldest to maintain bounded latency
-                                queue.pop_front();
-                                guard.record_frame_dropped();
-                                log::trace!("[Stream] Queue full, dropped oldest frame");
-                            }
-                            queue.push_back(frame);
-                        }
-                        Err(broadcast::error::RecvError::Lagged(n)) => {
-                            let now = TokioInstant::now();
-                            if last_lagged_log.map_or(true, |t| now.duration_since(t).as_secs() >= 1) {
-                                log::warn!("[Stream] Lagged by {n} frames");
-                                last_lagged_log = Some(now);
-                            }
-                            // Continue - next recv will get latest
-                        }
-                        Err(broadcast::error::RecvError::Closed) => {
-                            rx_closed = true;
-                            log::debug!("[Stream] Channel closed, draining {} queued frames", queue.len());
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Wrapper that logs HTTP audio stream lifecycle and tracks delivery timing.
-///
-/// Logs when the stream starts and ends, and tracks gaps in frame delivery
-/// to help diagnose issues where Sonos stops receiving audio unexpectedly.
-struct LoggingStreamGuard {
-    stream_id: String,
-    client_ip: IpAddr,
-    frames_sent: AtomicU64,
-    first_error: parking_lot::Mutex<Option<String>>,
-    /// Tracks delivery timing to detect gaps
-    delivery_stats: parking_lot::Mutex<DeliveryStats>,
-}
-
-/// Tracks frame delivery timing statistics.
-struct DeliveryStats {
-    last_delivery: Option<Instant>,
-    max_gap_ms: u64,
-    gaps_over_threshold: u64,
-    /// Number of times silence mode was entered
-    silence_events: u64,
-    /// Total silence frames injected
-    silence_frames: u64,
-    /// Frames dropped due to cadence queue overflow
-    frames_dropped: u64,
-}
-
-impl LoggingStreamGuard {
-    fn new(stream_id: String, client_ip: IpAddr) -> Self {
-        log::info!(
-            "[Stream] HTTP stream started: stream={}, client={}",
-            stream_id,
-            client_ip
-        );
-        Self {
-            stream_id,
-            client_ip,
-            frames_sent: AtomicU64::new(0),
-            first_error: parking_lot::Mutex::new(None),
-            delivery_stats: parking_lot::Mutex::new(DeliveryStats {
-                last_delivery: None,
-                max_gap_ms: 0,
-                gaps_over_threshold: 0,
-                silence_events: 0,
-                silence_frames: 0,
-                frames_dropped: 0,
-            }),
-        }
-    }
-
-    /// Records that silence mode was entered.
-    fn record_silence_event(&self) {
-        self.delivery_stats.lock().silence_events += 1;
-    }
-
-    /// Records a silence frame being sent.
-    fn record_silence_frame(&self) {
-        self.delivery_stats.lock().silence_frames += 1;
-    }
-
-    /// Records a frame being dropped due to queue overflow.
-    fn record_frame_dropped(&self) {
-        self.delivery_stats.lock().frames_dropped += 1;
-    }
-
-    fn record_frame(&self) {
-        self.frames_sent.fetch_add(1, Ordering::Relaxed);
-
-        let now = Instant::now();
-        let mut stats = self.delivery_stats.lock();
-
-        if let Some(last) = stats.last_delivery {
-            let gap_ms = now.duration_since(last).as_millis() as u64;
-
-            if gap_ms > stats.max_gap_ms {
-                stats.max_gap_ms = gap_ms;
-            }
-
-            if gap_ms > DELIVERY_GAP_THRESHOLD_MS {
-                stats.gaps_over_threshold += 1;
-                // Only log significant gaps to avoid spam; summary captures total count
-                if gap_ms > DELIVERY_GAP_LOG_THRESHOLD_MS {
-                    log::warn!(
-                        "[Stream] Delivery gap detected: stream={}, client={}, gap={}ms",
-                        self.stream_id,
-                        self.client_ip,
-                        gap_ms
-                    );
-                }
-            }
-        }
-
-        stats.last_delivery = Some(now);
-    }
-
-    fn record_error(&self, err: &str) {
-        let mut first = self.first_error.lock();
-        if first.is_none() {
-            *first = Some(err.to_string());
-        }
-    }
-}
-
-impl Drop for LoggingStreamGuard {
-    fn drop(&mut self) {
-        let frames = self.frames_sent.load(Ordering::Relaxed);
-        let first_error = self.first_error.get_mut();
-        let stats = self.delivery_stats.get_mut();
-
-        // Calculate time since last frame delivery
-        let final_gap_ms = stats
-            .last_delivery
-            .map(|t| t.elapsed().as_millis() as u64)
-            .unwrap_or(0);
-        let stalled_suffix = if final_gap_ms > DELIVERY_GAP_LOG_THRESHOLD_MS {
-            " (stalled)"
-        } else {
-            ""
-        };
-
-        // Build silence stats string if any silence was injected
-        let silence_info = if stats.silence_events > 0 {
-            format!(
-                ", silence_events={}, silence_frames={}",
-                stats.silence_events, stats.silence_frames
-            )
-        } else {
-            String::new()
-        };
-
-        // Build dropped frames string if any frames were dropped
-        let dropped_info = if stats.frames_dropped > 0 {
-            format!(", frames_dropped={}", stats.frames_dropped)
-        } else {
-            String::new()
-        };
-
-        if let Some(ref err) = *first_error {
-            log::warn!(
-                "[Stream] HTTP stream ended with error{}: stream={}, client={}, frames_sent={}, \
-                 max_gap={}ms, gaps_over_{}ms={}, final_gap={}ms{}{}, error={}",
-                stalled_suffix,
-                self.stream_id,
-                self.client_ip,
-                frames,
-                stats.max_gap_ms,
-                DELIVERY_GAP_THRESHOLD_MS,
-                stats.gaps_over_threshold,
-                final_gap_ms,
-                silence_info,
-                dropped_info,
-                err
-            );
-        } else {
-            log::info!(
-                "[Stream] HTTP stream ended normally{}: stream={}, client={}, frames_sent={}, \
-                 max_gap={}ms, gaps_over_{}ms={}, final_gap={}ms{}{}",
-                stalled_suffix,
-                self.stream_id,
-                self.client_ip,
-                frames,
-                stats.max_gap_ms,
-                DELIVERY_GAP_THRESHOLD_MS,
-                stats.gaps_over_threshold,
-                final_gap_ms,
-                silence_info,
-                dropped_info
-            );
-        }
-    }
-}
 
 // ─────────────────────────────────────────────────────────────────────────────
 // GENA Helpers
@@ -973,8 +586,8 @@ async fn stream_audio(
     // Type alias for tagged frame stream
     type TaggedStream = Pin<Box<dyn Stream<Item = Result<TaggedFrame, std::io::Error>> + Send>>;
 
-    // Create logging guard early so we can pass it to the cadence stream for drop tracking.
-    // Uses Arc so it can be shared between cadence stream, silence tracking, and final frame recording.
+    // Create logging guard early so we can pass it to the cadence stream for internal tracking.
+    // Uses Arc so it can be shared between cadence stream and final frame recording.
     let guard = Arc::new(LoggingStreamGuard::new(id.to_string(), remote_ip));
 
     // Build combined stream - PCM gets cadence-based streaming, compressed codecs don't.
@@ -1095,29 +708,8 @@ async fn stream_audio(
         builder = builder.header(header::CONTENT_LENGTH, WAV_STREAM_SIZE_MAX.to_string());
     }
 
-    // Unwrap TaggedFrame to Bytes, tracking silence events/frames for the final summary.
-    // This must happen before the TaggedFrame type information is lost.
-    let guard_for_silence = Arc::clone(&guard);
-    let mut silence_event_recorded = false;
-    let unwrapped_stream = tracked_stream.map(move |res| {
-        if let Ok(ref frame) = res {
-            match frame {
-                TaggedFrame::Silence(_) => {
-                    // Record first silence frame as an "event" (entering silence mode)
-                    if !silence_event_recorded {
-                        guard_for_silence.record_silence_event();
-                        silence_event_recorded = true;
-                    }
-                    guard_for_silence.record_silence_frame();
-                }
-                TaggedFrame::Audio(_) => {
-                    // Reset so next silence burst counts as new event
-                    silence_event_recorded = false;
-                }
-            }
-        }
-        res.map(TaggedFrame::into_bytes)
-    });
+    // Unwrap TaggedFrame to Bytes (silence tracking handled inside cadence stream)
+    let unwrapped_stream = tracked_stream.map(|res| res.map(TaggedFrame::into_bytes));
 
     // Apply ICY injection or PCM/WAV header to the unwrapped stream
     let inner_stream: AudioStream = if wants_icy {
@@ -1164,423 +756,6 @@ async fn stream_audio(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use futures::StreamExt;
-    use std::future::poll_fn;
-    use std::task::Poll;
-    use tokio::time::{self, Duration};
-
-    /// Test silence frame for assertions.
-    fn test_silence_frame() -> Bytes {
-        Bytes::from_static(&[0u8; 64])
-    }
-
-    /// Test audio frame for assertions.
-    fn test_audio_frame() -> Bytes {
-        Bytes::from_static(&[1u8; 64])
-    }
-
-    /// Polls the stream once to register internal timers, then advances time.
-    ///
-    /// With `start_paused = true`, timers must be polled before `time::advance`
-    /// will affect them. This helper ensures the stream's internal select! loop
-    /// registers its timers before we manipulate time.
-    async fn poll_and_advance<S>(stream: &mut Pin<&mut S>, duration: Duration)
-    where
-        S: Stream + ?Sized,
-    {
-        // Poll stream once to register timers (should return Pending since no data/timeout yet)
-        poll_fn(|cx| {
-            let _ = stream.as_mut().poll_next(cx);
-            Poll::Ready(())
-        })
-        .await;
-
-        time::advance(duration).await;
-    }
-
-    mod wav_cadence_streaming {
-        use super::*;
-        use crate::protocol_constants::SILENCE_FRAME_DURATION_MS;
-        use std::net::Ipv4Addr;
-
-        /// Default queue size for tests (10 frames = 200ms at 20ms/frame).
-        const TEST_QUEUE_SIZE: usize = 10;
-
-        /// Creates a test guard for cadence stream tests.
-        fn test_guard() -> Arc<LoggingStreamGuard> {
-            Arc::new(LoggingStreamGuard::new(
-                "test-stream".to_string(),
-                IpAddr::V4(Ipv4Addr::LOCALHOST),
-            ))
-        }
-
-        /// Creates a test audio format for cadence stream tests.
-        fn test_audio_format() -> AudioFormat {
-            AudioFormat::new(48000, 2, 16)
-        }
-
-        #[tokio::test(start_paused = true)]
-        async fn emits_frames_at_cadence() {
-            let (tx, rx) = broadcast::channel::<Bytes>(16);
-            let silence = test_silence_frame();
-            let audio = test_audio_frame();
-            let guard = test_guard();
-
-            let mut stream = Box::pin(create_wav_stream_with_cadence(
-                rx,
-                silence.clone(),
-                guard,
-                TEST_QUEUE_SIZE,
-                SILENCE_FRAME_DURATION_MS,
-                test_audio_format(),
-                vec![],
-            ));
-
-            // Queue some frames before the first tick
-            tx.send(audio.clone()).expect("send should succeed");
-            tx.send(audio.clone()).expect("send should succeed");
-
-            // Poll to register metronome, advance one tick (20ms)
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            // Should get first audio frame
-            let frame = stream.next().await.expect("stream should yield");
-            assert!(
-                matches!(frame.unwrap(), TaggedFrame::Audio(_)),
-                "expected first Audio frame at cadence tick"
-            );
-
-            // Advance another tick
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            // Should get second audio frame
-            let frame = stream.next().await.expect("stream should yield");
-            assert!(
-                matches!(frame.unwrap(), TaggedFrame::Audio(_)),
-                "expected second Audio frame at cadence tick"
-            );
-
-            drop(tx);
-        }
-
-        #[tokio::test(start_paused = true)]
-        async fn fills_gaps_with_silence() {
-            let (tx, rx) = broadcast::channel::<Bytes>(16);
-            let silence = test_silence_frame();
-            let guard = test_guard();
-
-            let mut stream = Box::pin(create_wav_stream_with_cadence(
-                rx,
-                silence.clone(),
-                guard,
-                TEST_QUEUE_SIZE,
-                SILENCE_FRAME_DURATION_MS,
-                test_audio_format(),
-                vec![],
-            ));
-
-            // Don't send any frames - queue will be empty
-
-            // Poll to register metronome, advance one tick
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            // Should get silence frame since queue is empty
-            let frame = stream.next().await.expect("stream should yield");
-            assert!(
-                matches!(frame.unwrap(), TaggedFrame::Silence(_)),
-                "expected Silence frame when queue is empty"
-            );
-
-            // Another tick should also yield silence
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            let frame = stream.next().await.expect("stream should yield");
-            assert!(
-                matches!(frame.unwrap(), TaggedFrame::Silence(_)),
-                "expected Silence frame on continued empty queue"
-            );
-
-            drop(tx);
-        }
-
-        #[tokio::test(start_paused = true)]
-        async fn queue_drains_at_cadence() {
-            let (tx, rx) = broadcast::channel::<Bytes>(16);
-            let silence = test_silence_frame();
-            let guard = test_guard();
-
-            let mut stream = Box::pin(create_wav_stream_with_cadence(
-                rx,
-                silence.clone(),
-                guard,
-                TEST_QUEUE_SIZE,
-                SILENCE_FRAME_DURATION_MS,
-                test_audio_format(),
-                vec![],
-            ));
-
-            // Queue 3 frames as a burst
-            for i in 0..3 {
-                tx.send(Bytes::from(vec![i; 64]))
-                    .expect("send should succeed");
-            }
-
-            // Each tick should drain one frame
-            for expected_byte in 0..3u8 {
-                poll_and_advance(
-                    &mut stream.as_mut(),
-                    Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-                )
-                .await;
-
-                let frame = stream.next().await.expect("stream should yield");
-                let frame = frame.expect("should be Ok");
-                if let TaggedFrame::Audio(bytes) = frame {
-                    assert_eq!(
-                        bytes[0], expected_byte,
-                        "frames should drain in order at cadence"
-                    );
-                } else {
-                    panic!("expected Audio frame, got Silence");
-                }
-            }
-
-            // Queue is now empty, next tick should yield silence
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            let frame = stream.next().await.expect("stream should yield");
-            assert!(
-                matches!(frame.unwrap(), TaggedFrame::Silence(_)),
-                "expected Silence after queue drained"
-            );
-
-            drop(tx);
-        }
-
-        #[tokio::test(start_paused = true)]
-        async fn drops_oldest_on_overflow() {
-            let (tx, rx) = broadcast::channel::<Bytes>(32);
-            let silence = test_silence_frame();
-            let guard = test_guard();
-            let guard_for_check = Arc::clone(&guard);
-
-            let mut stream = Box::pin(create_wav_stream_with_cadence(
-                rx,
-                silence.clone(),
-                guard,
-                TEST_QUEUE_SIZE,
-                SILENCE_FRAME_DURATION_MS,
-                test_audio_format(),
-                vec![],
-            ));
-
-            // Send 12 frames (queue_size + 2), should drop 2 oldest
-            for i in 0..12u8 {
-                tx.send(Bytes::from(vec![i; 64]))
-                    .expect("send should succeed");
-            }
-
-            // Poll multiple times to ensure frames are received via the rx.recv() branch
-            // and overflow logic is triggered
-            for _ in 0..15 {
-                poll_and_advance(
-                    &mut stream.as_mut(),
-                    Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-                )
-                .await;
-            }
-
-            // Verify that exactly 2 frames were dropped
-            let stats = guard_for_check.delivery_stats.lock();
-            assert_eq!(
-                stats.frames_dropped, 2,
-                "should have dropped 2 oldest frames, dropped {}",
-                stats.frames_dropped
-            );
-
-            drop(tx);
-        }
-
-        #[tokio::test(start_paused = true)]
-        async fn tracks_dropped_frames() {
-            let (tx, rx) = broadcast::channel::<Bytes>(32);
-            let silence = test_silence_frame();
-            let guard = test_guard();
-            let guard_for_check = Arc::clone(&guard);
-
-            let mut stream = Box::pin(create_wav_stream_with_cadence(
-                rx,
-                silence.clone(),
-                guard,
-                TEST_QUEUE_SIZE,
-                SILENCE_FRAME_DURATION_MS,
-                test_audio_format(),
-                vec![],
-            ));
-
-            // Overflow queue by TEST_QUEUE_SIZE + 3 frames
-            let overflow_count = 3;
-            for i in 0..(TEST_QUEUE_SIZE + overflow_count) as u8 {
-                tx.send(Bytes::from(vec![i; 64]))
-                    .expect("send should succeed");
-            }
-
-            // Give time for frames to be received and processed
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            // Consume a frame to ensure internal processing has happened
-            let _ = stream.next().await;
-
-            // Check the guard counter - should have recorded dropped frames
-            let stats = guard_for_check.delivery_stats.lock();
-            assert_eq!(
-                stats.frames_dropped, overflow_count as u64,
-                "guard should track {} dropped frames",
-                overflow_count
-            );
-
-            drop(tx);
-        }
-
-        #[tokio::test(start_paused = true)]
-        async fn drains_queue_on_channel_close() {
-            let (tx, rx) = broadcast::channel::<Bytes>(16);
-            let silence = test_silence_frame();
-            let guard = test_guard();
-
-            let mut stream = Box::pin(create_wav_stream_with_cadence(
-                rx,
-                silence.clone(),
-                guard,
-                TEST_QUEUE_SIZE,
-                SILENCE_FRAME_DURATION_MS,
-                test_audio_format(),
-                vec![],
-            ));
-
-            // Queue some frames
-            tx.send(Bytes::from(vec![1; 64]))
-                .expect("send should succeed");
-            tx.send(Bytes::from(vec![2; 64]))
-                .expect("send should succeed");
-
-            // Give time for frames to be queued
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-            let _ = stream.next().await; // consume first frame
-
-            // Close the channel
-            drop(tx);
-
-            // Advance and drain remaining frame
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            // Should still get the queued frame
-            let frame = stream
-                .next()
-                .await
-                .expect("stream should yield queued frame");
-            let frame = frame.expect("should be Ok");
-            if let TaggedFrame::Audio(bytes) = frame {
-                assert_eq!(bytes[0], 2, "should drain remaining queued frame");
-            } else {
-                panic!("expected Audio frame from queue");
-            }
-
-            // Now stream should end (channel closed AND queue empty)
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            let frame = stream.next().await;
-            assert!(
-                frame.is_none(),
-                "stream should end when channel closed and queue empty"
-            );
-        }
-
-        #[tokio::test(start_paused = true)]
-        async fn exits_silence_when_frames_arrive() {
-            let (tx, rx) = broadcast::channel::<Bytes>(16);
-            let silence = test_silence_frame();
-            let audio = test_audio_frame();
-            let guard = test_guard();
-
-            let mut stream = Box::pin(create_wav_stream_with_cadence(
-                rx,
-                silence.clone(),
-                guard,
-                TEST_QUEUE_SIZE,
-                SILENCE_FRAME_DURATION_MS,
-                test_audio_format(),
-                vec![],
-            ));
-
-            // First tick with empty queue -> silence
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            let frame = stream.next().await.expect("stream should yield");
-            assert!(
-                matches!(frame.unwrap(), TaggedFrame::Silence(_)),
-                "expected Silence when queue empty"
-            );
-
-            // Queue a frame
-            tx.send(audio.clone()).expect("send should succeed");
-
-            // Next tick should yield audio (exit silence)
-            poll_and_advance(
-                &mut stream.as_mut(),
-                Duration::from_millis(SILENCE_FRAME_DURATION_MS as u64),
-            )
-            .await;
-
-            let frame = stream.next().await.expect("stream should yield");
-            assert!(
-                matches!(frame.unwrap(), TaggedFrame::Audio(_)),
-                "expected Audio when frame queued after silence"
-            );
-
-            drop(tx);
-        }
-    }
 
     mod manual_speaker_handlers {
         use super::*;
